+++ packages/workspaces-api/src/shared/decoders.ts
@@ -410,7 +410,7 @@ export const streamRequestArgumentsDecoder: Decoder<{ type: WorkspaceEventType; 
     branch: nonEmptyStringDecoder
 });
 
-export const workspaceEventActionDecoder: Decoder<WorkspaceEventAction> = oneOf<"opened" | "closing" | "closed" | "focus" | "added" | "loaded" | "removed" | "childrenUpdate" | "containerChange" | "maximized" | "restored" | "minimized" | "normal" | "selected" | "lock-configuration-changed">(
+export const workspaceEventActionDecoder: Decoder<WorkspaceEventAction> = oneOf<"opened" | "closing" | "closed" | "focus" | "added" | "loaded" | "removed" | "childrenUpdate" | "containerChange" | "maximized" | "restored" | "minimized" | "normal" | "selected" | "lock-configuration-changed" | "hibernated" | "resumed">(
     constant("opened"),
     constant("closing"),
     constant("closed"),
@@ -425,7 +425,9 @@ export const workspaceEventActionDecoder: Decoder<WorkspaceEventAction> = oneOf<
     constant("minimized"),
     constant("normal"),
     constant("selected"),
-    constant("lock-configuration-changed")
+    constant("lock-configuration-changed"),
+    constant("hibernated"),
+    constant("resumed")
 );
 
 export const workspaceConfigResultDecoder: Decoder<WorkspaceConfigResult> = object({